Move prompt and text_to_recipe to app directory
